Moved 2 tests from test_type_utils to elementwise/test_type_utils
The dpctl/tests/test_type_utils.py has been removed.
